Flatten resource cleaner, have UninstallScheduler init with 0 expected resources
